+++ mapchete/formats/default/tile_directory.py
@@ -5,12 +5,12 @@ import logging
 import numpy as np
 import numpy.ma as ma
 import os
-from shapely.geometry import box
+from shapely.geometry import box, shape, mapping
 
 from mapchete.config import validate_values
 from mapchete.errors import MapcheteConfigError
 from mapchete.formats import base, load_output_writer
-from mapchete.io import path_exists, absolute_path, read_json
+from mapchete.io import path_exists, absolute_path, read_json, tile_to_zoom_level
 from mapchete.io.vector import reproject_geometry, read_vector_window
 from mapchete.io.raster import read_raster_window, create_mosaic, resample_from_array
 from mapchete.tile import BufferedTilePyramid
@@ -143,34 +143,93 @@ class InputData(base.InputData):
         else:
             self._profile = None
 
-    def open(self, tile, **kwargs):
+    def open(
+        self,
+        tile,
+        tile_directory_zoom=None,
+        matching_method="gdal",
+        matching_max_zoom=None,
+        matching_precision=8,
+        fallback_to_higher_zoom=False,
+        **kwargs
+    ):
         """
         Return InputTile object.
 
         Parameters
         ----------
         tile : ``Tile``
+        tile_directory_zoom : None
+            If set, data will be read from exactly this zoom level
+        matching_method : str ('gdal' or 'min') (default: 'gdal')
+            gdal: Uses GDAL's standard method. Here, the target resolution is calculated
+                by averaging the extent's pixel sizes over both x and y axes. This
+                approach returns a zoom level which may not have the best quality but will
+                speed up reading significantly.
+            min: Returns the zoom level which matches the minimum resolution of the
+                extents four corner pixels. This approach returns the zoom level with the
+                best possible quality but with low performance. If the tile extent is
+                outside of the destination pyramid, a TopologicalError will be raised.
+        matching_max_zoom : int (default: None)
+            If set, it will prevent reading from zoom levels above the maximum.
+        matching_precision : int
+            Round resolutions to n digits before comparing.
+        fallback_to_higher_zoom : bool (default: False)
+            In case no data is found at zoom level, try to read data from higher zoom
+            levels. Enabling this setting can lead to many IO requests in areas with no
+            data.
 
         Returns
         -------
         input tile : ``InputTile``
             tile view of input data
         """
+        # determine tile bounds in TileDirectory CRS
+        td_bounds = reproject_geometry(
+            tile.bbox,
+            src_crs=tile.tp.crs,
+            dst_crs=self.td_pyramid.crs
+        ).bounds
+
+        # find target zoom level
+        if tile_directory_zoom is not None:
+            zoom = tile_directory_zoom
+        else:
+            zoom = tile_to_zoom_level(
+                tile, dst_pyramid=self.td_pyramid, matching_method=matching_method,
+                precision=matching_precision
+            )
+            if matching_max_zoom is not None:
+                zoom = min([zoom, matching_max_zoom])
+
+        if fallback_to_higher_zoom:
+            tiles_paths = []
+            # check if tiles exist otherwise try higher zoom level
+            while len(tiles_paths) == 0 and zoom >= 0:
+                tiles_paths = _get_tiles_paths(
+                    basepath=self.path,
+                    ext=self._ext,
+                    pyramid=self.td_pyramid,
+                    bounds=td_bounds,
+                    zoom=zoom
+                )
+                logger.debug("%s existing tiles found at zoom %s", len(tiles_paths), zoom)
+                zoom -= 1
+        else:
+            tiles_paths = _get_tiles_paths(
+                basepath=self.path,
+                ext=self._ext,
+                pyramid=self.td_pyramid,
+                bounds=td_bounds,
+                zoom=zoom
+            )
+            logger.debug("%s existing tiles found at zoom %s", len(tiles_paths), zoom)
         return InputTile(
             tile,
-            tiles_paths=[
-                (_tile, _path)
-                for _tile, _path in [
-                    (t, os.path.join(*([
-                        self.path, str(t.zoom), str(t.row), str(t.col)
-                        ])) + "." + self._ext)
-                    for t in self.td_pyramid.tiles_from_bounds(
-                        tile.bounds, tile.zoom)
-                ]
-                if path_exists(_path)
-            ],
+            tiles_paths=tiles_paths,
             file_type=self._file_type,
             profile=self._profile,
+            td_crs=self.td_pyramid.crs,
             **kwargs
         )
 
@@ -195,6 +254,22 @@ class InputData(base.InputData):
         )
 
 
+def _get_tiles_paths(basepath=None, ext=None, pyramid=None, bounds=None, zoom=None):
+    return [
+        (_tile, _path)
+        for _tile, _path in [
+            (
+                t,
+                "%s.%s" % (
+                    os.path.join(*([basepath, str(t.zoom), str(t.row), str(t.col)])), ext
+                )
+            )
+            for t in pyramid.tiles_from_bounds(bounds, zoom)
+        ]
+        if path_exists(_path)
+    ]
+
+
 class InputTile(base.InputTile):
     """
     Target Tile representation of input data.
@@ -216,10 +291,15 @@ class InputTile(base.InputTile):
         self._tiles_paths = kwargs["tiles_paths"]
         self._file_type = kwargs["file_type"]
         self._profile = kwargs["profile"]
+        self._td_crs = kwargs["td_crs"]
 
     def read(
-        self, validity_check=False, indexes=None, resampling="nearest",
-        dst_nodata=None, gdal_opts=None
+        self,
+        validity_check=False,
+        indexes=None,
+        resampling="nearest",
+        dst_nodata=None,
+        gdal_opts=None
     ):
         """
         Read reprojected & resampled input data.
@@ -244,40 +324,37 @@ class InputTile(base.InputTile):
         -------
         data : list for vector files or numpy array for raster files
         """
+        logger.debug("reading data from CRS %s to CRS %s", self._td_crs, self.tile.tp.crs)
         if self._file_type == "vector":
             if self.is_empty():
                 return []
-            return list(chain.from_iterable([
-                read_vector_window(
-                    _path, self.tile, validity_check=validity_check)
-                for _, _path in self._tiles_paths
-            ]))
+            else:
+                return read_vector_window(
+                    [path for _, path in self._tiles_paths],
+                    self.tile,
+                    validity_check=validity_check
+                )
         else:
             if self.is_empty():
-                count = (len(indexes) if indexes else self._profile["count"], )
+                bands = len(indexes) if indexes else self._profile["count"]
                 return ma.masked_array(
                     data=np.full(
-                        count + self.tile.shape, self._profile["nodata"],
-                        dtype=self._profile["dtype"]),
+                        (bands, self.tile.height, self.tile.width),
+                        self._profile["nodata"],
+                        dtype=self._profile["dtype"]
+                    ),
                     mask=True
                 )
-            tiles = [
-                (
-                    _tile,
-                    read_raster_window(
-                        _path, _tile, indexes=indexes, resampling=resampling,
-                        src_nodata=self._profile["nodata"], dst_nodata=dst_nodata,
-                        gdal_opts=gdal_opts
-                    )
+            else:
+                return read_raster_window(
+                    [path for _, path in self._tiles_paths],
+                    self.tile,
+                    indexes=indexes,
+                    resampling=resampling,
+                    src_nodata=self._profile["nodata"],
+                    dst_nodata=dst_nodata,
+                    gdal_opts=gdal_opts
                 )
-                for _tile, _path in self._tiles_paths
-            ]
-            return resample_from_array(
-                in_raster=create_mosaic(tiles=tiles, nodata=self._profile["nodata"]),
-                out_tile=self.tile,
-                resampling=resampling,
-                nodataval=self._profile["nodata"]
-            )
 
     def is_empty(self):
         """