+++ mapchete/io/raster.py
@@ -21,12 +21,17 @@ from mapchete.io import path_is_remote, GDAL_HTTP_OPTS
 
 logger = logging.getLogger(__name__)
 
-ReferencedRaster = namedtuple("ReferencedRaster", ("data", "affine", "bounds"))
+ReferencedRaster = namedtuple("ReferencedRaster", ("data", "affine", "bounds", "crs"))
 
 
 def read_raster_window(
-    input_file, tile, indexes=None, resampling="nearest", src_nodata=None,
-    dst_nodata=None, gdal_opts=None
+    input_files,
+    tile,
+    indexes=None,
+    resampling="nearest",
+    src_nodata=None,
+    dst_nodata=None,
+    gdal_opts=None
 ):
     """
     Return NumPy arrays from an input raster.
@@ -38,8 +43,9 @@ def read_raster_window(
 
     Parameters
     ----------
-    input_file : string
-        path to a raster file readable by rasterio.
+    input_files : string or list
+        path to a raster file or list of paths to multiple raster files readable by
+        rasterio.
     tile : Tile
         a Tile object
     indexes : list or int
@@ -57,42 +63,90 @@ def read_raster_window(
     -------
     raster : MaskedArray
     """
-    dst_shape = tile.shape
+    input_file = input_files[0] if isinstance(input_files, list) else input_files
     user_opts = {} if gdal_opts is None else dict(**gdal_opts)
     if path_is_remote(input_file, s3=True):
         gdal_opts = dict(GDAL_HTTP_OPTS, **user_opts)
     else:
         gdal_opts = user_opts
-
-    if not isinstance(indexes, int):
-        if indexes is None:
-            dst_shape = (None,) + dst_shape
-        elif len(indexes) == 1:
-            indexes = indexes[0]
-        else:
-            dst_shape = (len(indexes),) + dst_shape
-    # Check if potentially tile boundaries exceed tile matrix boundaries on
-    # the antimeridian, the northern or the southern boundary.
-    if tile.pixelbuffer and tile.is_on_edge():
-        return _get_warped_edge_array(
-            tile=tile, input_file=input_file, indexes=indexes,
-            dst_shape=dst_shape, resampling=resampling, src_nodata=src_nodata,
-            dst_nodata=dst_nodata, gdal_opts=gdal_opts
+    with rasterio.Env(**gdal_opts):
+        return _read_raster_window(
+            input_files,
+            tile,
+            indexes=indexes,
+            resampling=resampling,
+            src_nodata=src_nodata,
+            dst_nodata=dst_nodata
         )
 
-    # If tile boundaries don't exceed pyramid boundaries, simply read window
-    # once.
+
+def _read_raster_window(
+    input_files,
+    tile,
+    indexes=None,
+    resampling="nearest",
+    src_nodata=None,
+    dst_nodata=None
+):
+    if isinstance(input_files, list):
+        # in case multiple input files are given, merge output into one array
+        # read first file
+        dst_array = _read_raster_window(
+            input_files[0],
+            tile=tile,
+            indexes=indexes,
+            resampling=resampling,
+            src_nodata=src_nodata,
+            dst_nodata=dst_nodata
+        )
+        # read subsequent files and merge
+        for f in input_files[1:]:
+            f_array = _read_raster_window(
+                f,
+                tile=tile,
+                indexes=indexes,
+                resampling=resampling,
+                src_nodata=src_nodata,
+                dst_nodata=dst_nodata
+            )
+            dst_array = ma.MaskedArray(
+                data=np.where(f_array.mask, dst_array, f_array).astype(dst_array.dtype),
+                mask=np.where(f_array.mask, dst_array.mask, f_array.mask).astype(np.bool)
+            )
+        return dst_array
     else:
-        return _get_warped_array(
-            input_file=input_file, indexes=indexes, dst_bounds=tile.bounds,
-            dst_shape=dst_shape, dst_crs=tile.crs, resampling=resampling,
-            src_nodata=src_nodata, dst_nodata=dst_nodata, gdal_opts=gdal_opts
-        )
+        input_file = input_files
+        dst_shape = tile.shape
+
+        if not isinstance(indexes, int):
+            if indexes is None:
+                dst_shape = (None,) + dst_shape
+            elif len(indexes) == 1:
+                indexes = indexes[0]
+            else:
+                dst_shape = (len(indexes),) + dst_shape
+        # Check if potentially tile boundaries exceed tile matrix boundaries on
+        # the antimeridian, the northern or the southern boundary.
+        if tile.pixelbuffer and tile.is_on_edge():
+            return _get_warped_edge_array(
+                tile=tile, input_file=input_file, indexes=indexes,
+                dst_shape=dst_shape, resampling=resampling, src_nodata=src_nodata,
+                dst_nodata=dst_nodata
+            )
+
+        # If tile boundaries don't exceed pyramid boundaries, simply read window
+        # once.
+        else:
+            return _get_warped_array(
+                input_file=input_file, indexes=indexes, dst_bounds=tile.bounds,
+                dst_shape=dst_shape, dst_crs=tile.crs, resampling=resampling,
+                src_nodata=src_nodata, dst_nodata=dst_nodata
+            )
 
 
 def _get_warped_edge_array(
     tile=None, input_file=None, indexes=None, dst_shape=None, resampling=None,
-    src_nodata=None, dst_nodata=None, gdal_opts=None
+    src_nodata=None, dst_nodata=None
 ):
     tile_boxes = clip_geometry_to_srs_bounds(
         tile.bbox, tile.tile_pyramid, multipart=True)
@@ -134,7 +188,7 @@ def _get_warped_edge_array(
             dst_bounds=parts_metadata[part]["bounds"],
             dst_shape=parts_metadata[part]["shape"],
             dst_crs=tile.crs, resampling=resampling, src_nodata=src_nodata,
-            dst_nodata=dst_nodata, gdal_opts=gdal_opts
+            dst_nodata=dst_nodata
         )
         for part in ["none", "left", "middle", "right"]
         if parts_metadata[part]
@@ -143,39 +197,37 @@ def _get_warped_edge_array(
 
 def _get_warped_array(
     input_file=None, indexes=None, dst_bounds=None, dst_shape=None,
-    dst_crs=None, resampling=None, src_nodata=None, dst_nodata=None,
-    gdal_opts=None
+    dst_crs=None, resampling=None, src_nodata=None, dst_nodata=None
 ):
     """Extract a numpy array from a raster file."""
     logger.debug("reading %s", input_file)
     try:
-        with rasterio.Env(**gdal_opts):
-            with rasterio.open(input_file, "r") as src:
-                if indexes is None:
-                    dst_shape = (len(src.indexes), dst_shape[-2], dst_shape[-1], )
-                    indexes = list(src.indexes)
-                src_nodata = src.nodata if src_nodata is None else src_nodata
-                dst_nodata = src.nodata if dst_nodata is None else dst_nodata
-                with WarpedVRT(
-                    src,
-                    crs=dst_crs,
-                    src_nodata=src_nodata,
-                    nodata=dst_nodata,
-                    width=dst_shape[-2],
-                    height=dst_shape[-1],
-                    transform=Affine(
-                        (dst_bounds[2] - dst_bounds[0]) / dst_shape[-2],
-                        0, dst_bounds[0], 0,
-                        (dst_bounds[1] - dst_bounds[3]) / dst_shape[-1],
-                        dst_bounds[3]
-                    ),
-                    resampling=Resampling[resampling]
-                ) as vrt:
-                    return vrt.read(
-                        window=vrt.window(*dst_bounds),
-                        out_shape=dst_shape,
-                        indexes=indexes,
-                        masked=True
+        with rasterio.open(input_file, "r") as src:
+            if indexes is None:
+                dst_shape = (len(src.indexes), dst_shape[-2], dst_shape[-1], )
+                indexes = list(src.indexes)
+            src_nodata = src.nodata if src_nodata is None else src_nodata
+            dst_nodata = src.nodata if dst_nodata is None else dst_nodata
+            with WarpedVRT(
+                src,
+                crs=dst_crs,
+                src_nodata=src_nodata,
+                nodata=dst_nodata,
+                width=dst_shape[-2],
+                height=dst_shape[-1],
+                transform=Affine(
+                    (dst_bounds[2] - dst_bounds[0]) / dst_shape[-2],
+                    0, dst_bounds[0], 0,
+                    (dst_bounds[1] - dst_bounds[3]) / dst_shape[-1],
+                    dst_bounds[3]
+                ),
+                resampling=Resampling[resampling]
+            ) as vrt:
+                return vrt.read(
+                    window=vrt.window(*dst_bounds),
+                    out_shape=dst_shape,
+                    indexes=indexes,
+                    masked=True
                     )
     except Exception as e:
         logger.exception("error while reading file %s: %s", input_file, e)
@@ -236,6 +288,7 @@ def write_raster_window(
     tags : optional tags to be added to GeoTIFF file
     bucket_resource : boto3 bucket resource to write to in case of S3 output
     """
+    logger.debug("write %s", out_path)
     if out_path == "memoryfile":
         raise DeprecationWarning(
             "Writing to memoryfile with write_raster_window() is deprecated. "
@@ -342,7 +395,11 @@ def extract_from_array(in_raster=None, in_affine=None, out_tile=None):
 
 
 def resample_from_array(
-    in_raster=None, in_affine=None, out_tile=None, resampling="nearest",
+    in_raster=None,
+    in_affine=None,
+    out_tile=None,
+    in_crs=None,
+    resampling="nearest",
     nodataval=0
 ):
     """
@@ -362,12 +419,14 @@ def resample_from_array(
     -------
     resampled array : array
     """
+    # TODO rename function
     if isinstance(in_raster, ma.MaskedArray):
         pass
     if isinstance(in_raster, np.ndarray):
         in_raster = ma.MaskedArray(in_raster, mask=in_raster == nodataval)
     elif isinstance(in_raster, ReferencedRaster):
         in_affine = in_raster.affine
+        in_crs = in_raster.crs
         in_raster = in_raster.data
     elif isinstance(in_raster, tuple):
         in_raster = ma.MaskedArray(
@@ -396,8 +455,12 @@ def resample_from_array(
         data=in_raster.filled(), mask=in_raster.mask, fill_value=nodataval
     )
     reproject(
-        in_raster, dst_data, src_transform=in_affine, src_crs=out_tile.crs,
-        dst_transform=out_tile.affine, dst_crs=out_tile.crs,
+        in_raster,
+        dst_data,
+        src_transform=in_affine,
+        src_crs=in_crs if in_crs else out_tile.crs,
+        dst_transform=out_tile.affine,
+        dst_crs=out_tile.crs,
         resampling=Resampling[resampling]
     )
     return ma.MaskedArray(dst_data, mask=dst_data == nodataval)
@@ -433,11 +496,16 @@ def create_mosaic(tiles, nodata=0):
     if len(tiles) == 0:
         raise ValueError("tiles list is empty")
 
-    logger.debug("mosaicking %s tile(s)", len(tiles))
+    logger.debug("create mosaic from %s tile(s)", len(tiles))
     # quick return if there is just one tile
     if len(tiles) == 1:
         tile, data = tiles[0]
-        return ReferencedRaster(data=data, affine=tile.affine, bounds=tile.bounds)
+        return ReferencedRaster(
+            data=data,
+            affine=tile.affine,
+            bounds=tile.bounds,
+            crs=tile.crs
+        )
 
     # assert all tiles have same properties
     pyramid, resolution, dtype = _get_tiles_properties(tiles)
@@ -493,7 +561,8 @@ def create_mosaic(tiles, nodata=0):
     return ReferencedRaster(
         data=mosaic,
         affine=affine,
-        bounds=Bounds(m_left, m_bottom, m_right, m_top)
+        bounds=Bounds(m_left, m_bottom, m_right, m_top),
+        crs=tile.crs
     )