+++ test/test_io.py
@@ -8,15 +8,19 @@ import tempfile
 import numpy as np
 import numpy.ma as ma
 import fiona
+from shapely.errors import TopologicalError
 from shapely.geometry import shape, box, Polygon, MultiPolygon
 from shapely.ops import unary_union
 from rasterio.enums import Compression
 from rasterio.crs import CRS
 from itertools import product
 
+import mapchete
 from mapchete.config import MapcheteConfig
 from mapchete.tile import BufferedTilePyramid
-from mapchete.io import get_best_zoom_level, path_exists, absolute_path, read_json
+from mapchete.io import (
+    get_best_zoom_level, path_exists, absolute_path, read_json, tile_to_zoom_level
+)
 from mapchete.io.raster import (
     read_raster_window, write_raster_window, extract_from_array,
     resample_from_array, create_mosaic, ReferencedRaster, prepare_array,
@@ -130,6 +134,37 @@ def test_read_raster_window_mask(s2_band):
         s2_band, tile, resampling="cubic", src_nodata=0, dst_nodata=0)
     assert data.any()
     assert not np.where(data == 1, True, False).any()
+
+
+def test_read_raster_window_input_list(cleantopo_br):
+    process_zoom = 5
+    conf = dict(**cleantopo_br.dict)
+    conf["output"].update(metatiling=1)
+    with mapchete.open(conf) as mp:
+        mp.batch_process(process_zoom)
+        tiles = [
+            (tile, mp.config.output.get_path(tile))
+            for tile in mp.config.output_pyramid.tiles_from_bounds(
+                mp.config.bounds, process_zoom
+            )
+            if path_exists(mp.config.output.get_path(tile))
+        ]
+        upper_tile = next(mp.get_process_tiles(process_zoom - 1))
+        assert len(tiles) > 1
+        resampled = resample_from_array(
+            in_raster=create_mosaic(
+                [(tile, read_raster_window(path, tile)) for tile, path in tiles]
+            ),
+            out_tile=upper_tile
+        )
+    resampled2 = read_raster_window(
+        [p for _, p in tiles], upper_tile, src_nodata=0, dst_nodata=0
+    )
+    assert resampled.dtype == resampled2.dtype
+    assert resampled.shape == resampled2.shape
+    assert np.array_equal(resampled.mask, resampled2.mask)
+    # TODO slight rounding errors occur
+    assert np.allclose(resampled, resampled2, rtol=0.01)
 
 
 def test_write_raster_window():
@@ -682,5 +717,82 @@ def test_read_json(s3_metadata_json, http_metadata_json):
         read_json("https://ungarj.github.io/mapchete_testdata/tiled_data/raster/cleantopo/invalid_metadata.json")
 
 
-# TODO write_vector_window()
-# TODO extract_from_tile()
+def test_tile_to_zoom_level():
+    tp_merc = BufferedTilePyramid("mercator")
+    tp_geod = BufferedTilePyramid("geodetic")
+    zoom = 9
+    col = 0
+
+    # mercator from geodetic
+    # at Northern boundary
+    assert tile_to_zoom_level(
+        tp_merc.tile(zoom, 0, col),
+        tp_geod
+    ) == 9
+    assert tile_to_zoom_level(
+        tp_merc.tile(zoom, 0, col),
+        tp_geod,
+        matching_method="min"
+    ) == 12
+    # at Equator
+    assert tile_to_zoom_level(
+        tp_merc.tile(zoom, tp_merc.matrix_height(zoom) // 2, col),
+        tp_geod
+    ) == 9
+    assert tile_to_zoom_level(
+        tp_merc.tile(zoom, tp_merc.matrix_height(zoom) // 2, col),
+        tp_geod,
+        matching_method="min"
+    ) == 9
+    # at Southern boundary
+    assert tile_to_zoom_level(
+        tp_merc.tile(zoom, tp_merc.matrix_height(zoom) - 1, col),
+        tp_geod
+    ) == 9
+    assert tile_to_zoom_level(
+        tp_merc.tile(zoom, tp_merc.matrix_height(zoom) - 1, col),
+        tp_geod,
+        matching_method="min"
+    ) == 12
+
+    # geodetic from mercator
+    # at Northern boundary
+    assert tile_to_zoom_level(
+        tp_geod.tile(zoom, 0, col),
+        tp_merc
+    ) == 2
+    with pytest.raises(TopologicalError):
+        tile_to_zoom_level(
+            tp_geod.tile(zoom, 0, col),
+            tp_merc,
+            matching_method="min"
+        )
+    # at Equator
+    assert tile_to_zoom_level(
+        tp_geod.tile(zoom, tp_geod.matrix_height(zoom) // 2, col),
+        tp_merc
+    ) == 10
+    assert tile_to_zoom_level(
+        tp_geod.tile(zoom, tp_geod.matrix_height(zoom) // 2, col),
+        tp_merc,
+        matching_method="min"
+    ) == 10
+    # at Southern boundary
+    assert tile_to_zoom_level(
+        tp_geod.tile(zoom, tp_geod.matrix_height(zoom) - 1, col),
+        tp_merc
+    ) == 2
+    with pytest.raises(TopologicalError):
+        tile_to_zoom_level(
+            tp_geod.tile(zoom, tp_geod.matrix_height(zoom) - 1, col),
+            tp_merc,
+            matching_method="min"
+        )
+
+    # check wrong method
+    with pytest.raises(ValueError):
+        tile_to_zoom_level(
+            tp_geod.tile(zoom, tp_geod.matrix_height(zoom) - 1, col),
+            tp_merc,
+            matching_method="invalid_method"
+        )